+++ Kernel/CMakeLists.txt
@@ -276,7 +276,7 @@ set(KERNEL_SOURCES
     Net/Intel/E1000NetworkAdapter.cpp
     Net/Realtek/RTL8168NetworkAdapter.cpp
     Net/VirtIO/VirtIONetworkAdapter.cpp
-    Net/IPv4/Socket.cpp
+    Net/IP/Socket.cpp
     Net/LocalSocket.cpp
     Net/LoopbackAdapter.cpp
     Net/NetworkAdapter.cpp
