+++ Kernel/FileSystem/SysFS/Subsystems/Kernel/Network/ARP.cpp
@@ -6,7 +6,7 @@
 
 #include <AK/JsonObjectSerializer.h>
 #include <Kernel/FileSystem/SysFS/Subsystems/Kernel/Network/ARP.h>
-#include <Kernel/Net/IPv4/ARP.h>
+#include <Kernel/Net/IP/ARP.h>
 #include <Kernel/Net/Routing.h>
 #include <Kernel/Sections.h>
