+++ Kernel/Net/IP/Socket.cpp
@@ -12,10 +12,10 @@
 #include <Kernel/Debug.h>
 #include <Kernel/FileSystem/OpenFileDescription.h>
 #include <Kernel/Net/ICMP.h>
-#include <Kernel/Net/IPv4/ARP.h>
-#include <Kernel/Net/IPv4/IP.h>
-#include <Kernel/Net/IPv4/IPv4.h>
-#include <Kernel/Net/IPv4/Socket.h>
+#include <Kernel/Net/IP/ARP.h>
+#include <Kernel/Net/IP/IP.h>
+#include <Kernel/Net/IP/IPv4.h>
+#include <Kernel/Net/IP/Socket.h>
 #include <Kernel/Net/NetworkAdapter.h>
 #include <Kernel/Net/NetworkingManagement.h>
 #include <Kernel/Net/Routing.h>
